// src/types.ts
/**
 * @var CoreTypes
 * @description Core types for dependency system.
 */
export enum CoreTypes {
    EventBus = '$eventBus',
    ModuleStore = '$moduleStore',
    ResourceStore = '$resourceStore',
    InternationalizationStore = '$internationalizationStore',
    ConfigStore = '$configStore',
    ModuleService = '$moduleService',
    ResourceService = '$resourceService',
    InternationalizationService = '$internationalizationService',
    ConfigService = '$configService',
}